WIP Adding in the basic structure of handling connections/config
