+++ database/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "database"
+version = "0.1.0"
+authors = ["Shem Sedrick <ssedrick1@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+syntax = { path = "../syntax" }
+net = { path = "../net" }
+clap = { version = "2", features = ["yaml"] }
+tokio = { version = "0.2", features = ["full"] }
