+++ net/Cargo.toml
@@ -1,10 +1,8 @@
 [package]
-name = "syntax"
+name = "net"
 version = "0.1.0"
 authors = ["Shem Sedrick <ssedrick1@gmail.com>"]
 edition = "2018"
 
 [dependencies]
-lazy_static = "1"
-regex = "1"
-log = "*"
+
