+++ net/src/lib.rs
@@ -1,0 +1,9 @@
+pub mod handlers;
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn it_works() {
+        assert_eq!(2 + 2, 4);
+    }
+}
